Moved archive toggle component to filter components section - This makes more sense than having it in the projects components
